# openapi.yaml
components:
  schemas:
...
    BatchSearchRequest:
      description: Request for batch search.
      properties:
        channel:
          default: latest
          description: Version channel to search
          title: Channel
          type: string
        entity_type:
          default: memory_card
          description: Type of entity to search
          title: Entity Type
          type: string
        hybrid_weights:
          default:
          - 0.5
          - 0.5
          maxItems: 2
          minItems: 2
          prefixItems:
          - type: number
          - type: number
          title: Hybrid Weights
          type: array
        namespace:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by namespace
          title: Namespace
        queries:
          description: List of search queries
          items:
            type: string
          maxItems: 50
          minItems: 1
          title: Queries
          type: array
        query_vectors:
          anyOf:
          - items:
              items:
                type: number
              type: array
            type: array
          - type: 'null'
          description: Pre-computed embedding vectors for vector search
          title: Query Vectors
        search_type:
          $ref: '#/components/schemas/SearchType'
          default: bm25
          description: Type of search
        tags:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          description: Filter by tags
          title: Tags
        top_k:
          default: 10
          description: Number of top results per query
          maximum: 100.0
          minimum: 1.0
          title: Top K
          type: integer
      required:
      - queries
      title: BatchSearchRequest
      type: object
    BatchSearchResponseModel:
      description: Response for batch search.
      properties:
        results:
          items:
            items:
              $ref: '#/components/schemas/SearchHit'
            type: array
          title: Results
          type: array
        search_type:
          $ref: '#/components/schemas/SearchType'
        total_queries:
          title: Total Queries
          type: integer
      required:
      - results
      - search_type
      - total_queries
      title: BatchSearchResponseModel
      type: object
...
    EntityCreateRequest:
      properties:
        channel:
          default: latest
          title: Channel
          type: string
        content:
          additionalProperties: true
          title: Content
          type: object
        embedding:
          anyOf:
          - items:
              type: number
            type: array
          - type: 'null'
          title: Embedding
        evolution_meta:
          anyOf:
          - $ref: '#/components/schemas/EvolutionMeta'
          - type: 'null'
        meta:
          $ref: '#/components/schemas/EntityMeta'
        parent_version_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Parent Version Id
      required:
      - meta
      - content
      title: EntityCreateRequest
      type: object
    EntityMeta:
      properties:
        author:
          anyOf:
          - type: string
          - type: 'null'
          title: Author
        name:
          maxLength: 255
          minLength: 1
          title: Name
          type: string
        namespace:
          anyOf:
          - type: string
          - type: 'null'
          title: Namespace
        tags:
          items:
            type: string
          title: Tags
          type: array
        when_to_use:
          anyOf:
          - type: string
          - type: 'null'
          title: When To Use
      required:
      - name
      title: EntityMeta
      type: object
...
    EntityUpdateRequest:
      properties:
        change_summary:
          anyOf:
          - type: string
          - type: 'null'
          title: Change Summary
        channel:
          default: latest
          title: Channel
          type: string
        content:
          additionalProperties: true
          title: Content
          type: object
        embedding:
          anyOf:
          - items:
              type: number
            type: array
          - type: 'null'
          title: Embedding
        evolution_meta:
          anyOf:
          - $ref: '#/components/schemas/EvolutionMeta'
          - type: 'null'
        meta:
          anyOf:
          - $ref: '#/components/schemas/EntityMeta'
          - type: 'null'
        parent_version_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Parent Version Id
      required:
      - content
      title: EntityUpdateRequest
      type: object
    EvolutionMeta:
      properties:
        behavioral_descriptors:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Behavioral Descriptors
        fitness:
          anyOf:
          - type: number
          - type: 'null'
          title: Fitness
        is_valid:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Valid
        metrics:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Metrics
        mutation_kind:
          anyOf:
          - type: string
          - type: 'null'
          title: Mutation Kind
        prompt_ref:
          anyOf:
          - type: string
          - type: 'null'
          title: Prompt Ref
      title: EvolutionMeta
      type: object
    FacetsResponse:
      properties:
        authors:
          additionalProperties:
            type: integer
          title: Authors
          type: object
        entity_types:
          additionalProperties:
            type: integer
          title: Entity Types
          type: object
        namespaces:
          additionalProperties:
            type: integer
          title: Namespaces
          type: object
        tags:
          additionalProperties:
            type: integer
          title: Tags
          type: object
      title: FacetsResponse
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    SearchHit:
      properties:
        channel:
          anyOf:
          - type: string
          - type: 'null'
          title: Channel
        content:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Content
        entity_id:
          title: Entity Id
          type: string
        entity_type:
          title: Entity Type
          type: string
        name:
          title: Name
          type: string
        score:
          default: 0.0
          title: Score
          type: number
        tags:
          items:
            type: string
          title: Tags
          type: array
        version_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Version Id
        when_to_use:
          anyOf:
          - type: string
          - type: 'null'
          title: When To Use
      required:
      - entity_id
      - entity_type
      - name
      title: SearchHit
      type: object
    SearchType:
      description: Search algorithm type.
      enum:
      - bm25
      - vector
      - hybrid
      title: SearchType
      type: string
    StepResponse:
      properties:
        channel:
          title: Channel
          type: string
        content:
          additionalProperties: true
          title: Content
          type: object
        entity_id:
          title: Entity Id
          type: string
        entity_type:
          const: step
          default: step
          title: Entity Type
          type: string
        etag:
          title: Etag
          type: string
        meta:
          additionalProperties: true
          title: Meta
          type: object
        version_id:
          title: Version Id
          type: string
      required:
      - entity_id
      - version_id
      - channel
      - etag
      - meta
      - content
      title: StepResponse
      type: object
    UnifiedSearchRequest:
      description: Request for unified search (BM25, vector, or hybrid).
      properties:
        channel:
          default: latest
          description: Version channel to search
          title: Channel
          type: string
        entity_type:
          default: memory_card
          description: Type of entity to search
          title: Entity Type
          type: string
        hybrid_weights:
          default:
          - 0.5
          - 0.5
          description: (bm25_weight, vector_weight) for hybrid search, must sum to
            1.0
          maxItems: 2
          minItems: 2
          prefixItems:
          - type: number
          - type: number
          title: Hybrid Weights
          type: array
        namespace:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by namespace
          title: Namespace
        query:
          anyOf:
          - type: string
          - type: 'null'
          description: Text query for BM25 or hybrid search
          title: Query
        query_vector:
          anyOf:
          - items:
              type: number
            type: array
          - type: 'null'
          description: Pre-computed embedding vector for vector search
          title: Query Vector
        search_type:
          $ref: '#/components/schemas/SearchType'
          default: bm25
          description: 'Type of search: ''bm25'', ''vector'', or ''hybrid'''
        tags:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          description: Filter by tags
          title: Tags
        top_k:
          default: 10
          description: Number of top results to return
          maximum: 100.0
          minimum: 1.0
          title: Top K
          type: integer
      title: UnifiedSearchRequest
      type: object
    UnifiedSearchResponse:
      description: Response for unified search (BM25 or vector).
      properties:
        hits:
          items:
            $ref: '#/components/schemas/SearchHit'
          title: Hits
          type: array
        search_type:
          description: Type of search performed (bm25 or vector)
          title: Search Type
          type: string
        total:
          description: Total number of hits returned
          title: Total
          type: integer
      required:
      - search_type
      - total
      title: UnifiedSearchResponse
      type: object
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
info:
  description: 'Persistent memory for CARL artifacts: steps, chains, agents, memory
    cards'
  title: GigaEvo Memory Module
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /v1/search/batch:
    post:
      description: "Batch search for multiple queries.\n\nEfficiently processes multiple\
        \ queries in parallel using asyncio.gather.\nReturns a list of result lists,\
        \ one per query.\n\nExample request:\n    {\n        \"search_type\": \"hybrid\"\
        ,\n        \"queries\": [\"financial analysis\", \"code review\", \"documentation\"\
        ],\n        \"top_k\": 5,\n        \"entity_type\": \"memory_card\",\n   \
        \     \"hybrid_weights\": [0.4, 0.6]\n    }"
      operationId: batch_search_v1_search_batch_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchSearchRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchSearchResponseModel'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Batch Search
      tags:
      - search
  /v1/search/facets:
    get:
      description: 'Get aggregated facet counts for UI filters.


        Returns counts for:

        - entity_types: Number of entities per type

        - tags: Number of entities per tag

        - authors: Number of entities per author

        - namespaces: Number of entities per namespace'
      operationId: get_facets_v1_search_facets_get
      parameters:
      - description: Filter facets by namespace
        in: query
        name: namespace
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter facets by namespace
          title: Namespace
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FacetsResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Facets
      tags:
      - search
  /v1/search/unified:
    post:
      description: "Unified search endpoint supporting BM25, vector, and hybrid search.\n\
        \nThis endpoint provides a single interface for three search types:\n- **bm25**:\
        \ Full-text search using PostgreSQL tsvector\n- **vector**: Semantic search\
        \ using pgvector similarity\n- **hybrid**: Combined BM25 + Vector scores with\
        \ configurable weights\n\nFor vector and hybrid search, embeddings can be\
        \ generated server-side\n(if ENABLE_VECTOR_SEARCH=true) or provided as pre-computed\
        \ vectors.\n\nExample BM25 request:\n    {\n        \"search_type\": \"bm25\"\
        ,\n        \"query\": \"financial analysis\",\n        \"top_k\": 10,\n  \
        \      \"entity_type\": \"memory_card\"\n    }\n\nExample vector request (server-side\
        \ embedding):\n    {\n        \"search_type\": \"vector\",\n        \"query\"\
        : \"semantic similarity\",\n        \"top_k\": 10,\n        \"entity_type\"\
        : \"memory_card\"\n    }\n\nExample hybrid request:\n    {\n        \"search_type\"\
        : \"hybrid\",\n        \"query\": \"analysis\",\n        \"hybrid_weights\"\
        : [0.3, 0.7],\n        \"top_k\": 10,\n        \"entity_type\": \"memory_card\"\
        \n    }\n\nExample vector request (client-side embedding):\n    {\n      \
        \  \"search_type\": \"vector\",\n        \"query_vector\": [0.1, 0.2, ...],\n\
        \        \"top_k\": 10,\n        \"entity_type\": \"memory_card\"\n    }"
      operationId: unified_search_v1_search_unified_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnifiedSearchRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnifiedSearchResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Unified Search
      tags:
      - search
  /v1/steps:
    get:
      description: List all steps with pagination.
      operationId: list_steps_v1_steps_get
      parameters:
      - in: query
        name: limit
        required: false
        schema:
          default: 50
          title: Limit
          type: integer
      - in: query
        name: offset
        required: false
        schema:
          default: 0
          title: Offset
          type: integer
      - in: query
        name: channel
        required: false
        schema:
          default: latest
          title: Channel
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/StepResponse'
                title: Response List Steps V1 Steps Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List Steps
      tags:
      - steps
    post:
      description: Create a new step entity with its first version.
      operationId: create_step_v1_steps_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EntityCreateRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StepResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Step
      tags:
      - steps
  /v1/steps/{step_id}:
    delete:
      description: Soft-delete a step.
      operationId: delete_step_v1_steps__step_id__delete
      parameters:
      - in: path
        name: step_id
        required: true
        schema:
          format: uuid
          title: Step Id
          type: string
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete Step
      tags:
      - steps
    get:
      description: Get a step by ID, resolving the channel to a specific version.
      operationId: get_step_v1_steps__step_id__get
      parameters:
      - in: path
        name: step_id
        required: true
        schema:
          format: uuid
          title: Step Id
          type: string
      - in: query
        name: channel
        required: false
        schema:
          default: latest
          title: Channel
          type: string
      - in: header
        name: if-none-match
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: If-None-Match
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StepResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Step
      tags:
      - steps
    put:
      description: Update a step by creating a new immutable version.
      operationId: update_step_v1_steps__step_id__put
      parameters:
      - in: path
        name: step_id
        required: true
        schema:
          format: uuid
          title: Step Id
          type: string
      - in: header
        name: if-match
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: If-Match
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EntityUpdateRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StepResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Update Step
      tags:
      - steps
...